Move some remaining wrapper creator related code into the relevant folder
